Master: C++11 recognition realized by cmake, osx build now works, fixed bunch of warnings on MSVC

+++ monitors/onlinemon/include/TGraphSet.hh
@@ -27,7 +27,7 @@ protected:
   double _yLimitMax;
   double _yLimitMin;
 
-  float _starttime;
+  double _starttime;
   bool _nodata;
   TGraph *GetTGraph(std::string name);
   std::string GetUnit(std::string name);
